KP-6837: Move LdapAdapterTest.java to package
